ml_pipeline/skew_job/requirements.txt
# Skew Detection Job Dependencies
# ================================

# Google Cloud
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.14.0
google-cloud-storage>=2.14.0
earthengine-api>=0.1.380
db-dtypes>=1.2.0

# ML & Data Science
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization (for TensorBoard figures)
matplotlib>=3.8.0
seaborn>=0.13.0

# TensorBoard (PyTorch - matches trainer/)
tensorboard>=2.14.0
torch>=2.0.0  # For SummaryWriter

# Web Server
flask>=3.0.0
gunicorn>=21.2.0

# Config
pyyaml>=6.0.1
